+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:get_storage/get_storage.dart';
 import 'model/constants.dart';
 import 'screens/signing_screen.dart';
-import 'screens/user_profile_screen.dart';
+import 'screens/userProfile/user_profile_screen.dart';
 import 'screens/welcome_screen.dart';
 import 'screens/scanner_screen.dart';
 import 'screens/loading_screen.dart';
